Ajustes geral e refatoraçao
