+++ src/app/components/cadastro/cadastro.component.css
@@ -1,11 +1,3 @@
-.header-bottom {
-  display: flex;
-  align-items: center;
-  justify-content: start;
-  gap: 1px;
-  background-color: var(--primary-bg);
-}
-
 /* Bloco de Tecnologia */
 .bloco-tecnologia {
   display: flex;
@@ -19,28 +11,6 @@
   color: var(--text-primary);
   border-right: 1px solid var(--border-contrast);
   border-bottom: 1px solid var(--border-contrast);
-}
-
-.bloco-tecnologia .conteudo {
-  display: flex;
-  flex-direction: column;
-  gap: 1px;
-  width: 296px;
-}
-
-.bloco-tecnologia .titulo {
-  font-weight: var(--font-weight-light);
-  font-size: var(--font-normal);
-  line-height: 125%;
-  color: var(--text-muted);
-}
-
-.bloco-tecnologia .status {
-  display: flex;
-  width: 86px;
-  height: 29px;
-  font-weight: var(--font-weight-bold);
-  font-size: 22px; 
 }
 
 /* ====================== */
@@ -314,26 +284,6 @@
   border-radius: var(--border-radius-sm);
   font-weight: var(--font-weight-bold);
   text-align: center;
-}
-
-.score.verde {
-  color: #2ecc71;
-  background: rgba(46, 204, 113, 0.1);
-}
-
-.score.amarelo {
-  color: #f1c40f;
-  background: rgba(241, 196, 15, 0.1);
-}
-
-.score.vermelho-claro {
-  color: #e67e22;
-  background: rgba(230, 126, 34, 0.1);
-}
-
-.score.vermelho-escuro {
-  color: #e74c3c;
-  background: rgba(231, 76, 60, 0.1);
 }
 
 /* Ícones de ordenação */
