+++ src/app/components/sucessoras/sucessoras.component.css
@@ -708,34 +708,6 @@ input, button {
     }
 }  
 
-.score.verde {
-    color: #2ecc71;
-    background: rgba(46, 204, 113, 0.1);
-    padding: 4px 10px;
-    border-radius: 8px;
-  }
-  
-  .score.amarelo {
-    color: #f1c40f;
-    background: rgba(241, 196, 15, 0.1);
-    padding: 4px 10px;
-    border-radius: 8px;
-  }
-  
-  .score.vermelho-claro {
-    color: #e67e22;
-    background: rgba(230, 126, 34, 0.1);
-    padding: 4px 10px;
-    border-radius: 8px;
-  }
-  
-  .score.vermelho-escuro {
-    color: #e74c3c;
-    background: rgba(231, 76, 60, 0.1);
-    padding: 4px 10px;
-    border-radius: 8px;
-  }
-
   .link-titulo {
       font-size: 1.25rem;
       font-weight: 600;
