+++ src/app/components/tarefas/tarefas.component.css
@@ -1,11 +1,3 @@
-.header-bottom {
-    display: flex;
-    align-items: center;
-    justify-content: start;
-    gap: 1px;
-    background-color: #223348;
-}
-
 /* Bloco de Tecnologia */
 .bloco-tecnologia {
     display: flex;
@@ -14,16 +6,10 @@
     justify-content: center;
     min-width: 320px;
     height: 148px;
-    padding: 32px;
+    padding: 32px !important;
     color: #f5f9ff;
     border-right: 1px solid rgba(173, 200, 235, 0.14);
     border-bottom: 1px solid rgba(173, 200, 235, 0.14);
-}
-
-.bloco-tecnologia .conteudo {
-    display: flex;
-    flex-direction: column;
-    width: 296px;
 }
 
 /* Breadcrumb */
@@ -33,14 +19,6 @@
     line-height: 125%;
     color: rgba(245, 249, 255, 0.7); /* Cor com transparência */
     display: block; /* Garante que o breadcrumb fique em sua própria linha */
-}
-
-/* Título */
-.bloco-tecnologia .titulo {
-    font-weight: 600;
-    font-size: 22px;
-    line-height: 125%;
-    color: white;
 }
 
 /* Status */
@@ -104,85 +82,71 @@
   
 /* Bloco de Indicadores */
 .bloco-indicadores {
-    position: relative;
-    display: flex;
-    flex-direction: row;
-    min-width: 320px;
-    height: 148px;
-    padding: 16px;
-    color: #f5f9ff;
-    text-align: center;
-    border-right: 1px solid rgba(173, 200, 235, 0.14);
-    border-bottom: 1px solid rgba(173, 200, 235, 0.14);
+  position: relative;
+  display: flex;
+  min-width: 320px;
+  height: 148px;
+  color: #f5f9ff;
+  gap: 86px;
+  border-right: 1px solid rgba(173, 200, 235, 0.14);
+  border-bottom: 1px solid rgba(173, 200, 235, 0.14);
 }
 
 .bloco-indicadores::after {
-    content: "";
-    display: block;
-    width: 100%;
-    height: 5px;
-    position: absolute;
-    bottom: 0;
-    left: 0;
-    background: linear-gradient(to right, #0C6CE9 50%, #E0241A 50%);
+  content: "";
+  display: block;
+  width: 100%;
+  height: 5px;
+  position: absolute;
+  bottom: 0;
+  left: 0;
+  background: linear-gradient(to right,
+      #0C6CE9 0%,
+      #0C6CE9 33.33%,
+      #E0241A 33.33%,
+      #E0241A 66.66%,
+      #FFD600 66.66%,
+      #FFD600 100%);
 }
 
-.bloco-indicadores::before {
-    content: "";
-    display: block;
-    width: 1px;
-    height: calc(100% - 5px);
-    position: absolute;
-    top: 0;
-    left: 50%;
-    transform: translateX(-50%);
-    background-color: rgba(173, 200, 235, 0.14);
+/* Separadores verticais */
+.bloco-indicadores .divider {
+  position: absolute;
+  top: 0;
+  bottom: 0;
+  width: 1px;
+  background: rgba(173, 200, 235, 0.2);
+  z-index: 2;
 }
 
-.bloco-indicadores .pendente,
-.bloco-indicadores .atrasadas {
-    display: flex;
-    flex-direction: column; /* 🔹 Organiza os itens na vertical */
-    align-items: flex-start; /* 🔹 Alinha os elementos à esquerda */
-    justify-content: center;
-    gap: 8px;
-    width: 180px;
-    height: 100%;
-    padding: 16px;
+/* Posição dos separadores: 33.33% e 66.66% */
+.bloco-indicadores .divider.div1 {
+  left: calc(33.33%);
+  /* entre Pendente e Atrasadas */
 }
 
-.bloco-indicadores svg {
-    width: 20px;
-    height: 20px;
-    color: #f5f9ff;
-    margin-bottom: 8px; /* 🔹 Dá um pequeno espaço entre o ícone e o texto */
-
+.bloco-indicadores .divider.div2 {
+  left: calc(66.66%);
+  /* entre Atrasadas e Aguardando */
 }
 
-.texto-numero {
-    display: flex;
-    flex-direction: column;
-    align-items: flex-start; /* 🔹 Alinha os textos à esquerda */
-    justify-content: center;
-    width: 104px;
-    height: 54px;
-    gap: 4px;
+.numero-interativo {
+  cursor: pointer;
+  transition:
+    background 0.2s,
+    color 0.2s,
+    box-shadow 0.2s,
+    transform 0.2s;
+  /* adiciona animação suave ao scale */
+  border-radius: 6px;
 }
 
-.texto-numero .texto {
-    font-weight: 600;
-    font-size: 13px;
-    line-height: 125%;
-    letter-spacing: 3%;
-    text-transform: uppercase;
-    color: #f5f9ff;
-}
-
-.texto-numero .numero {
-    font-weight: 600;
-    font-size: 40px;
-    line-height: 125%;
-    color: #f5f9ff;
+.numero-interativo:hover {
+  background: #243447;
+  color: #fff;
+  box-shadow: 0 2px 8px rgba(0, 0, 0, 0.12);
+  transform: scale(1.10);
+  /* aumenta o tamanho em 10% */
 }
 
 /* ====================== */
@@ -254,7 +218,7 @@
   
   /* Campo de texto */
   .input-text {
-    min-width: 596px;
+    min-width: 320px;
     width: 100%;
     height: var(--height-small);
     padding: var(--spacing-xs);
