+++ src/app/components/usuarios/usuarios.component.css
@@ -1,11 +1,3 @@
-.header-bottom {
-  display: flex;
-  align-items: center;
-  justify-content: start;
-  gap: 1px;
-  background-color: var(--primary-bg);
-}
-
 /* Bloco de Tecnologia */
 .bloco-tecnologia {
   display: flex;
@@ -19,28 +11,6 @@
   color: var(--text-primary);
   border-right: 1px solid var(--border-contrast);
   border-bottom: 1px solid var(--border-contrast);
-}
-
-.bloco-tecnologia .conteudo {
-  display: flex;
-  flex-direction: column;
-  gap: 1px;
-  width: 296px;
-}
-
-.bloco-tecnologia .titulo {
-  font-weight: var(--font-weight-light);
-  font-size: var(--font-normal);
-  line-height: 125%;
-  color: var(--text-muted);
-}
-
-.bloco-tecnologia .status {
-  display: flex;
-  width: 86px;
-  height: 29px;
-  font-weight: var(--font-weight-bold);
-  font-size: 22px; /* Manter fixo, pois não tem variável */
 }
 
 /* ====================== */
