+++ src/environments/environment.prod.ts
@@ -1,4 +1,4 @@
 export const environment = {
   production: true,
-  backendApiUrl: 'http://192.168.10.3:3000/api'
-  };
+  backendApiUrl: 'https://legacy.jallcard.com.br/api'
+};
