+++ src/styles.css
@@ -65,6 +65,151 @@
   --modal-padding: 20px;
 }
 
+.header-bottom {
+  display: flex;
+  align-items: center;
+  justify-content: start;
+  gap: 1px;
+  background-color: var(--primary-bg);
+}
+
+.bloco-tecnologia .status {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  padding: 4px 8px;
+  width: 86px;
+  /* Mantém o tamanho correto */
+  height: 29px;
+  border-radius: 8px;
+  background: var(--semantic-colors-success-10, #0FF56F1A);
+  color: #0FF56F;
+  font-weight: 500;
+  font-size: 14px;
+}
+
+.bloco-tecnologia .titulo {
+  font-weight: 600;
+  font-size: 22px;
+  line-height: 125%;
+  color: white;
+}
+
+.bloco-tecnologia .conteudo {
+  display: flex;
+  flex-direction: column;
+  width: 296px;
+}
+
+/* Bloco de Indicadores */
+.bloco-indicadores .pendente,
+.bloco-indicadores .atrasadas,
+.bloco-indicadores .aguardando-setor {
+  flex: 1 1 0;
+  min-width: 0;
+  height: 100%;
+  padding: 30px 0;
+  /* só padding em cima/baixo */
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  justify-content: center;
+  gap: 8px;
+  box-sizing: border-box;
+}
+
+.bloco-indicadores svg {
+  width: 20px;
+  height: 20px;
+  color: #f5f9ff;
+  margin-bottom: 8px;
+  /* 🔹 Dá um pequeno espaço entre o ícone e o texto */
+}
+
+.texto-numero {
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  /* 🔹 Alinha os textos à esquerda */
+  justify-content: center;
+  width: 104px;
+  height: 54px;
+  gap: 4px;
+}
+
+.texto-numero .texto {
+  font-weight: 600;
+  font-size: 13px;
+  line-height: 125%;
+  letter-spacing: 3%;
+  text-transform: uppercase;
+  color: #f5f9ff;
+}
+
+.texto-numero .numero {
+  font-weight: 600;
+  font-size: 40px;
+  line-height: 125%;
+  color: #f5f9ff;
+}
+
+/* Status bolinha */
+.status-bolinha {
+  display: inline-block;
+  width: 10px;
+  height: 10px;
+  border-radius: 50%;
+  margin-right: 6px;
+  vertical-align: middle;
+  border: 1px solid #bbb;
+}
+
+.status-bolinha.verde {
+  background: #18aa61;
+  border-color: #18aa61;
+}
+
+.status-bolinha.vermelha {
+  background: #E0241A;
+  border-color: #E0241A;
+}
+
+.status-bolinha.amarela {
+  background: #FFD600;
+  border-color: #FFD600;
+}
+
+/* ====================== */
+/* SCORES */
+/* ====================== */
+.score.verde {
+  color: #2ecc71;
+  background: rgba(46, 204, 113, 0.1);
+  padding: 4px 10px;
+  border-radius: 8px;
+}
+
+.score.amarelo {
+  color: #f1c40f;
+  background: rgba(241, 196, 15, 0.1);
+  padding: 4px 10px;
+  border-radius: 8px;
+}
+
+.score.vermelho-claro {
+  color: #e67e22;
+  background: rgba(230, 126, 34, 0.1);
+  padding: 4px 10px;
+  border-radius: 8px;
+}
+
+.score.vermelho-escuro {
+  color: #e74c3c;
+  background: rgba(231, 76, 60, 0.1);
+  padding: 4px 10px;
+  border-radius: 8px;
+}
+
 /* Reset básico */
 * {
   margin: 0;
@@ -75,4 +220,45 @@
 body {
   font-family: 'Inter', sans-serif;
   background-color: #2B3D52; /* Fundo geral da página */
+}
+
+@media (max-width: 900px) {
+
+  .bloco-tecnologia,
+  .bloco-indicadores,
+  .bloco-filtro,
+  .quadro {
+    min-width: 220px;
+    padding: var(--spacing-sm);
+  }
+
+  .tabela-tarefas th,
+  .tabela-tarefas td {
+    font-size: var(--font-small);
+    padding: 6px;
+  }
+}
+
+@media (min-width: 1440px) {
+
+  .bloco-tecnologia,
+  .bloco-indicadores,
+  .bloco-filtro {
+    padding: var(--spacing-lg);
+  }
+
+  .titulo {
+    font-size: 26px;
+  }
+
+  .texto-numero .numero {
+    font-size: 48px;
+  }
+}
+
+/* ====================== */
+/* OUTROS AJUSTES */
+/* ====================== */
+.shadow-hover:hover {
+  box-shadow: 0 0 10px rgba(0, 0, 0, 0.5);
 }
